requirements.txt
discord.py
aiohttp
dotenv